Remade handleOnClick on RooferCard
